requirements.txt
pandas
SQLAlchemy
requests
psycopg2-binary
apache-airflow-providers-postgres
finnhub-python